requirements.txt
google-cloud-pubsub==2.1.0
google-cloud-storage==1.31.2
google-cloud-bigquery==2.1.0
google-cloud-core==1.4.3
google-auth
google-oauth
requests
pytz==2018.7
six